+++ include/bolt/cl/detail/scan_by_key.inl
@@ -295,9 +295,9 @@ size_t k0_stepNum, k1_stepNum, k2_stepNum;
      *********************************************************************************/
     bool cpuDevice = ctl.getDevice().getInfo<CL_DEVICE_TYPE>() == CL_DEVICE_TYPE_CPU;
     //std::cout << "Device is CPU: " << (cpuDevice?"TRUE":"FALSE") << std::endl;
-    const size_t kernel0_WgSize = (cpuDevice) ? 1 : WAVESIZE*KERNEL02WAVES;
-    const size_t kernel1_WgSize = (cpuDevice) ? 1 : WAVESIZE*KERNEL1WAVES;
-    const size_t kernel2_WgSize = (cpuDevice) ? 1 : WAVESIZE*KERNEL02WAVES;
+    const int kernel0_WgSize = (cpuDevice) ? 1 : WAVESIZE*KERNEL02WAVES;
+    const int kernel1_WgSize = (cpuDevice) ? 1 : WAVESIZE*KERNEL1WAVES;
+    const int kernel2_WgSize = (cpuDevice) ? 1 : WAVESIZE*KERNEL02WAVES;
     std::string compileOptions;
     std::ostringstream oss;
     oss << " -DKERNEL0WORKGROUPSIZE=" << kernel0_WgSize;
@@ -330,7 +330,7 @@ size_t k0_stepNum, k1_stepNum, k2_stepNum;
     cl_uint numElements = static_cast< cl_uint >( std::distance( firstKey, lastKey ) );
     typename device_vector< kType >::size_type sizeInputBuff = numElements;
 
-    size_t modWgSize = (sizeInputBuff & ((kernel0_WgSize*2)-1));
+    int modWgSize = (sizeInputBuff & ((kernel0_WgSize*2)-1));
     if( modWgSize )
     {
         sizeInputBuff &= ~modWgSize;
